requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.5
contourpy==1.2.1
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2025.3.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
idna==3.7
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.2.0
protobuf==4.25.3
Pygments==2.17.2
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.90
pywin32==306
requests==2.31.0
rich==13.7.1
scipy==1.13.0
setuptools==69.5.1
six==1.16.0
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.1.3
wheel==0.43.0
wrapt==1.16.0
